+++ cryptol-remote-api/python/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for `cryptol` Python package
 
+## next -- TBA
+
+Nothing yet.
+
 ## 3.3.0 -- 2025-03-21
 
 * Add `modules` command to return the modules, submodules and associated