+++ cryptol-remote-api/python/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for `cryptol` Python package
 
-## next
+## 3.3.0 -- 2025-03-21
 
 * Add `modules` command to return the modules, submodules and associated
   documentation for them.